+++ test/helpers/searchHelpers.js
